Fixed: dash

+++ scenes/main/bash_point_controller.gd
@@ -43,7 +43,7 @@ func create_shape(p):
 	for point in current_points:
 		if point == points[0]:
 			break
-		elif point.has_method("delete_point"):
+		elif point and point.has_method("delete_point"):
 				point.delete_point()
 	current_shape.finish_shape(points)
 	current_shape = null
